app/requirements.txt
requests
openai
tiktoken
youtube-transcript-api
python_dotenv
python-telegram-bot
